+++ hello-web/components/header.tsx
@@ -1,0 +1,11 @@
+import SessionManage from "@/components/session-manage";
+
+export default async function Header() {
+
+  return (
+    <div>
+      <SessionManage />
+      <hr />
+    </div>
+  )
+}
